init tags share

+++ web/src/lib/components/asset-viewer/detail-panel-tags.svelte
@@ -37,7 +37,7 @@
 
 <svelte:document use:shortcut={{ shortcut: { key: 't' }, onShortcut: handleAddTag }} />
 
-{#if isOwner && !authManager.key}
+{#if !authManager.key}
   <section class="px-4 mt-4">
     <div class="flex h-10 w-full items-center justify-between text-sm">
       <h2>{$t('tags').toUpperCase()}</h2>
@@ -53,27 +53,30 @@
               {tag.value}
             </p>
           </a>
-
-          <button
-            type="button"
-            class="text-gray-100 dark:text-immich-dark-gray bg-immich-primary/95 dark:bg-immich-dark-primary/95 rounded-e-full place-items-center place-content-center pe-2 ps-1 py-1 hover:bg-immich-primary/80 dark:hover:bg-immich-dark-primary/80 transition-all"
-            title="Remove tag"
-            onclick={() => handleRemove(tag.id)}
-          >
-            <Icon path={mdiClose} />
-          </button>
+          {#if isOwner && !authManager.key}
+            <button
+              type="button"
+              class="text-gray-100 dark:text-immich-dark-gray bg-immich-primary/95 dark:bg-immich-dark-primary/95 rounded-e-full place-items-center place-content-center pe-2 ps-1 py-1 hover:bg-immich-primary/80 dark:hover:bg-immich-dark-primary/80 transition-all"
+              title="Remove tag"
+              onclick={() => handleRemove(tag.id)}
+            >
+              <Icon path={mdiClose} />
+            </button>
+          {/if}
         </div>
       {/each}
-      <button
-        type="button"
-        class="rounded-full bg-gray-100 dark:bg-gray-800 text-gray-600 dark:text-gray-300 hover:bg-gray-200 dark:hover:bg-gray-700 hover:text-gray-700 dark:hover:text-gray-200 flex place-items-center place-content-center gap-1 px-2 py-1"
-        title="Add tag"
-        onclick={handleAddTag}
-      >
-        <span class="text-sm px-1 flex place-items-center place-content-center gap-1"
-          ><Icon path={mdiPlus} />{$t('add')}</span
+      {#if isOwner && !authManager.key}
+        <button
+          type="button"
+          class="rounded-full bg-gray-100 dark:bg-gray-800 text-gray-600 dark:text-gray-300 hover:bg-gray-200 dark:hover:bg-gray-700 hover:text-gray-700 dark:hover:text-gray-200 flex place-items-center place-content-center gap-1 px-2 py-1"
+          title="Add tag"
+          onclick={handleAddTag}
         >
-      </button>
+          <span class="text-sm px-1 flex place-items-center place-content-center gap-1"
+            ><Icon path={mdiPlus} />{$t('add')}</span
+          >
+        </button>
+      {/if}
     </section>
   </section>
 {/if}
